fix: move function to utils
